# server/openapi.yaml
openapi: 3.1.0
info:
  title: FilmRoom Fantasy Football API
  version: 1.0.0
  description: |
    Backend API for FilmRoom Fantasy Football. Built on Cloudflare Workers with Hono, D1 (SQLite), and Drizzle ORM.

    ## Authentication
    Most endpoints require a Bearer token in the `Authorization` header:
    ```
    Authorization: Bearer <jwt_token>
    ```
    Tokens are obtained via `/api/auth/login`, `/api/auth/register`, or `/api/auth/google`.
    Tokens expire after 24 hours. Sessions can be revoked via `/api/auth/logout`.

    ## Rate Limiting
    Auth endpoints are rate limited to 10 requests per 15 minutes per IP.
    Rate limit headers are returned on all rate-limited endpoints:
    - `X-RateLimit-Limit` — max requests in window
    - `X-RateLimit-Remaining` — remaining requests
    - `X-RateLimit-Reset` — Unix timestamp when window resets
    - `Retry-After` — seconds to wait (on 429 only)

    ## Admin Endpoints
    Admin sync endpoints require `X-Admin-Key` header matching the `SYNC_SECRET` environment variable.
  contact:
    name: FilmRoom
  license:
    name: MIT
servers:
  - url: http://localhost:8787
    description: Local development
  - url: https://filmroom-api.your-domain.workers.dev
    description: Production
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    AdminKey:
      type: apiKey
      in: header
      name: X-Admin-Key
...
paths:
...
  /api/players:
    get:
      tags: [Players]
      summary: Search & filter NFL players
      parameters:
        - name: q
          in: query
          schema: { type: string }
          description: Search by name
        - name: position
          in: query
          schema: { type: string, enum: [QB, RB, WR, TE, K, DEF] }
        - name: team
          in: query
          schema: { type: string }
        - name: status
          in: query
          schema: { type: string }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: limit
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        '200':
          description: Paginated player list
...
  /api/leagues:
    get:
      tags: [Leagues]
      summary: Get user's leagues
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: List of leagues
    post:
      tags: [Leagues]
      summary: Create a new league
      security: [{ BearerAuth: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                scoringFormat: { type: string, enum: [ppr, half-ppr, standard] }
                teamCount: { type: integer }
                seasonYear: { type: integer }
      responses:
        '201':
          description: League created
...
  /api/admin/sync-players:
    post:
      tags: [Admin]
      summary: Sync NFL players from Sleeper API
      security: [{ AdminKey: [] }]
      responses:
        '200':
          description: Sync results with insert/update counts
  /api/admin/sync-stats:
    post:
      tags: [Admin]
      summary: Sync weekly player stats from Sleeper
      security: [{ AdminKey: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                seasonYear: { type: integer }
                maxWeeks: { type: integer, minimum: 1, maximum: 22 }
      responses:
        '200':
          description: Stats sync results
  /api/admin/sync-projections:
    post:
      tags: [Admin]
      summary: Sync player projections from Sleeper
      security: [{ AdminKey: [] }]
      responses:
        '200':
          description: Projection sync results
  /api/admin/sync-games:
    post:
      tags: [Admin]
      summary: Sync NFL games from ESPN
      security: [{ AdminKey: [] }]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                seasonYear: { type: integer, minimum: 2000, maximum: 2100 }
                weeks: { type: array, items: { type: integer, minimum: 1, maximum: 22 } }
      responses:
        '200':
          description: Games sync results
  /api/admin/sync-news:
    post:
      tags: [Admin]
      summary: Sync player injury news from Sleeper
      security: [{ AdminKey: [] }]
      responses:
        '200':
          description: News sync results
  /api/admin/sync-twitter-news:
    post:
      tags: [Admin]
      summary: Sync sports news from RSS feeds
      security: [{ AdminKey: [] }]
      responses:
        '200':
          description: RSS news sync results

  # ─── YAHOO ────────────────────
...
  /api/yahoo/leagues:
    get:
      tags: [Yahoo]
      summary: Get Yahoo Fantasy leagues
      security: [{ BearerAuth: [] }]
      responses:
        '200':
          description: Yahoo Fantasy NFL leagues
...
  /api/feedback:
    post:
      tags: [Feedback]
      summary: Submit user feedback
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [type, message]
              properties:
                type: { type: string, enum: [bug, feature, general] }
                message: { type: string }
                email: { type: string, format: email }
                page: { type: string }
      responses:
        '201':
          description: Feedback submitted